// app/types/ProductTypes.tsx
export interface ProductType {
  id: Number;
  slug: string;
  nameBn: string;
  category: string;
  categoryNameBn: string;
  categoryIcon: string;
  unit: string;
  image: string;
  today: number;
  yesterday: number;
  lastWeek: number;
  lastMonth: number;
  change: {
    dir: string;
    pct: number;
  };
  markets: [
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
    {
      market: string;
      division: string;
      min: number;
      max: number;
    },
  ];
}